refactor: remove cycle dependencies and add some structure - some tests are updated
