+++ lib/handler/attachersHandler.js
@@ -1,12 +1,11 @@
 import {
-  connectElements,
   DEFAULT_CELL_HEIGHT,
   DEFAULT_CELL_WIDTH,
+  connectElements,
+  getMid,
   getBounds,
   getDockingPoint,
-  getMid
-} from '../utils/layoutUtil.js';
-import { getAttachedOutgoingElements } from '../utils/elementUtils.js';
+} from '../controllers/layoutController/utils.js'
 
 export default {
   'createElementDi': ({ element, row, col, diFactory }) => {
