+++ lib/handler/elementHandler.js
@@ -1,5 +1,5 @@
 import { is } from '../di/DiUtil.js';
-import { getBounds } from '../utils/layoutUtil.js';
+import { getBounds } from '../controllers/layoutController/utils.js';
 
 export default {
   'createElementDi': ({ element, row, col, diFactory }) => {
