+++ lib/handler/outgoingHandler.js
@@ -1,18 +1,25 @@
-import { connectElements } from '../utils/layoutUtil.js';
 import { is } from '../di/DiUtil.js';
 import {
   findElementInTree,
-  getAttachedOutgoingElements, getHostIncoming,
+  getAttachedOutgoingElements,
   getIncomingElements,
-  getOutgoingElements, isBoundaryEvent, getBoundaryIncomingHosts
+  getOutgoingElements,
 } from '../utils/elementUtils.js';
+
 import {
   getPathSegments,
+  getDirection,
   getPositionVerticalCrossedBy,
   getPositionsHorizontalCrossedBy,
   getCrossGrid,
-  getDirection,
-} from "../Layouter.js";
+  connectElements,
+  getOutgoingFromHost,
+  getOutgoingFromBoundary,
+  getIncomingFromHost,
+  getIncomingFromBoundaryHost,
+  sortElementsTopLeftBottomRight,
+  sortByType,
+} from '../controllers/layoutController/utils.js'
 
 export default {
   'addToGrid': ({ element, grid, stack }) => {
@@ -28,8 +35,6 @@ export default {
       // Здесь разрыв !!!!
       // return [element]
     }
-
-
 
     // получаем новые которых нет в гриде
     const newOutgoing = getNewOutgoing(element, grid)
@@ -97,13 +102,7 @@ export default {
 
 // helpers /////
 
-export function sortByType(arr, type) {
-  const nonMatching = arr.filter(item => !is(item, type));
-  const matching = arr.filter(item => is(item, type));
 
-  return [ ...matching, ...nonMatching ];
-  // return [ ...nonMatching, ...matching ];
-}
 
 function checkForLoop(element, visited, reverse) {
 
@@ -367,44 +366,9 @@ function getNewOutgoing(element, grid) {
   // return sortByType(processingElements, 'bpmn:Task')
 }
 
-export function sortElementsTopLeftBottomRight (grid) {
-  return function (a, b){
-    const aPos = grid.find(a);
-    const bPos = grid.find(b);
 
-    if (aPos && !bPos) return -1;
-    if (!aPos && bPos) return 1;
-    if (!aPos && !bPos) return 0;
 
-    return aPos[0] - bPos[0] || aPos[1] - bPos[1];
-  }
-}
 
-export function sortElementsTopLeftBottomRightColumn (grid) {
-  return function (a, b){
-    const aPos = grid.find(a);
-    const bPos = grid.find(b);
-
-    if (aPos && !bPos) return -1;
-    if (!aPos && bPos) return 1;
-    if (!aPos && !bPos) return 0;
-
-    return  aPos[1] - bPos[1] || aPos[0] - bPos[0];
-  }
-}
-
-export function sortColsLeftRightRowsBottomTop (grid) {
-  return function (a, b){
-    const aPos = grid.find(a);
-    const bPos = grid.find(b);
-
-    if (aPos && !bPos) return -1;
-    if (!aPos && bPos) return 1;
-    if (!aPos && !bPos) return 0;
-
-    return  aPos[1] - bPos[1] || bPos[0] - aPos[0];
-  }
-}
 
 function getInsertPosition (element, previousElement, grid, isBoundarySource) {
 
@@ -703,203 +667,7 @@ function pushVerticalEdgeBy(elements, grid){
   }
 }
 
-export function getVerticalChain(element, grid, oldChain) {
-  const chain = !oldChain ? new Set() : oldChain
-  if (!element) return chain;
-
-  const elementPosition = grid.find(element)
-  if (!elementPosition) return chain;
-
-  chain.add(element)
-
-  // схема подразумевает только связи с направлением N_S или S_N если path.length > 0
-
-  // что то здесь!!!!!
-  const edges = getExistElementEdges(element, grid)
-      .filter(edge => {
-        const {direction, path, sourceIsBoundary, sourcePosition, targetPosition} = edge
-        // для всех случаем N_S
-        if (direction === 'N_S') return true;
-        // S_N только если не идем в обход
-        // if (direction === 'S_N' && sourcePosition[0] === targetPosition[0] + 1  ) return true;
-        if (direction === 'S_N' && sourcePosition[0] === targetPosition[0] + 1  ) return true;
-        if (direction === 'S_N') {
-          // проверяем наличие элементов в диапазоне
-          const inRange = []
-          for (let rowIndex = sourcePosition[0] - 1; rowIndex > targetPosition; rowIndex--) {
-            const candidate = grid.get(rowIndex, sourcePosition[1])
-            if (candidate) inRange.push(candidate)
-          }
-          return inRange.length === 0
-        }
-        // if (direction === 'NW_SE' && sourceIsBoundary && sourcePosition[0] === targetPosition[0] - 1 && sourcePosition[1] === targetPosition[1] - 1) return true;
-        return false;
-      })
-
-  for (const edge of edges){
-    const nextElement = edge.source === element ? edge.target : edge.source
-    if (!chain.has(nextElement)) {
-      const nextChain = getVerticalChain(nextElement, grid, chain)
-      for (const nextChainEl of nextChain) {
-        chain.add(nextChainEl)
-      }
-    }
-  }
-  return chain
-}
-
-export function getHorizontalChain(element, grid, oldChain) {
-  const chain = !oldChain ? new Set() : oldChain
-  if (!element) return chain;
-
-  const elementPosition = grid.find(element)
-  if (!elementPosition) return chain;
-
-  chain.add(element)
-
-  // схема подразумевает только связи с направлением N_S или S_N если path.length > 0
-  const edges = getExistElementEdges(element, grid)
-      .filter(edge => {
-        const {direction, path, sourceIsBoundary, sourcePosition, targetPosition} = edge
-        // Для случаев W_E только если path.length > 0
-        // if (direction === 'W_E' && path.length > 0) return true;
-        if (direction === 'W_E') return true;
-        // E_W только если не идем в обход
-        // Todo: походу косяк в схеме вопрос как двигаться назад по горизонтали если никого нет по середине
-        // пока так
-        if (direction === 'E_W' && path.length > 0) return true;
-        // добавляем boundary ребра в диагонали
-        // if (direction === 'NW_SE' && sourceIsBoundary && sourcePosition[0] === targetPosition[0] - 1 && sourcePosition[1] === targetPosition[1] - 1) return true;
-
-        return false;
-      })
-
-  for (const edge of edges){
-    const nextElement = edge.source === element ? edge.target : edge.source
-    if (!chain.has(nextElement)) {
-      const nextChain = getHorizontalChain(nextElement, grid, chain)
-      for (const nextChainEl of nextChain) {
-        chain.add(nextChainEl)
-      }
-    }
-  }
-  return chain
-}
 
 
-// Возвращает исходящие только из элемента
-export function getOutgoingFromHost (element, isFlipped) {
-  return !isFlipped ? new Set(getOutgoingElements(element)) : new Set(getHostIncoming (element))
-}
 
-// Возвращает исходящие только из boundary элемента
-export function getOutgoingFromBoundary (element, isFlipped) {
-  return !isFlipped ? new Set (getAttachedOutgoingElements(element)) : new Set()
-}
 
-// Возвращает входящие только в элемент
-export function getIncomingFromHost (element, isFlipped) {
-  return !isFlipped ? new Set(getHostIncoming (element)) : new Set(getOutgoingElements(element))
-}
-
-// Возвращает входящие только в boundary
-export function getIncomingFromBoundaryHost(element, isFlipped) {
-  return !isFlipped ? new Set (getBoundaryIncomingHosts (element)) : new Set ()
-}
-
-/**
- * Возвращает ребро
- * source: Element
- * sourcePosition: [number, number]
- * sourceIsBoundary: boolean
- * target: Element
- * targetPosition: [number, number]
- * targetIsBoundary: boolean
- * direction: Direction
- * path: PathSegment[]
-
- *
- * @param element
- * @param isFlipped
- */
-export function getExistElementEdges(element, grid) {
-  const edges = []
-
-  const elementPosition = grid.find(element)
-  if (!elementPosition) return edges
-
-  //получаем исходящие из хоста
-  // Todo: сделать сортировку?
-  const outgoingFromHost = [...getOutgoingFromHost (element, grid.isFlipped)]
-      .filter(item => grid.hasElement(item))
-  for (const outgoing of outgoingFromHost) {
-    const sourcePosition = grid.find(element)
-    const targetPosition = grid.find(outgoing)
-    const edge = {source: element,sourcePosition, target: outgoing, targetPosition}
-    const outgoingPosition = grid.find(outgoing)
-    if (!outgoingPosition) continue;
-    const direction = getDirection(elementPosition, outgoingPosition)
-    edge.direction = direction
-
-    // Todo: объект пути это направление, путь и возможно тип
-    const path = getPathSegments(element, outgoing, grid)
-    edge.path = path
-    edges.push(edge)
-  }
-
-  // получаем исходящие из boundary
-  const outgoingFromBoundary = [...getOutgoingFromBoundary (element, grid.isFlipped)]
-      .filter(item => grid.hasElement(item))
-  for (const outgoing of outgoingFromBoundary) {
-    const sourcePosition = grid.find(element)
-    const targetPosition = grid.find(outgoing)
-    const edge = {source: element, sourcePosition, target: outgoing, targetPosition, sourceIsBoundary: true}
-    const outgoingPosition = grid.find(outgoing)
-    if (!outgoingPosition) continue;
-    const direction = getDirection(elementPosition, outgoingPosition)
-    edge.direction = direction
-
-    // Todo: объект пути это направление, путь и возможно тип
-    const path = getPathSegments(element, outgoing, grid, true)
-    edge.path = path
-    edges.push(edge)
-  }
-
-  // входящие из host
-  const incomingFromHost = [...getIncomingFromHost (element, grid.isFlipped)]
-      .filter(item => grid.hasElement(item))
-  for (const incoming of incomingFromHost) {
-    const sourcePosition = grid.find(incoming)
-    const targetPosition = grid.find(element)
-    const edge = {source: incoming, sourcePosition, target: element, targetPosition}
-    const incomingPosition = grid.find(incoming)
-    if (!incomingPosition) continue;
-    const direction = getDirection(incomingPosition, elementPosition)
-    edge.direction = direction
-
-    // Todo: объект пути это направление, путь и возможно тип
-    const path = getPathSegments(incoming, element, grid)
-    edge.path = path
-    edges.push(edge)
-  }
-
-  // входящие из boundary
-  const incomingFromBoundaryHost = [...getIncomingFromBoundaryHost(element, grid.isFlipped)]
-      .filter(item => grid.hasElement(item))
-  for (const incoming of incomingFromBoundaryHost) {
-    const sourcePosition = grid.find(incoming)
-    const targetPosition = grid.find(element)
-    const edge = {source: incoming, sourcePosition, target: element, targetPosition, sourceIsBoundary: true}
-    // const edge = {source: incoming, target: element, sourceIsBoundary: true}
-    const incomingPosition = grid.find(incoming)
-    if (!incomingPosition) continue;
-    const direction = getDirection(incomingPosition, elementPosition)
-    edge.direction = direction
-
-    // Todo: объект пути это направление, путь и возможно тип
-    const path = getPathSegments(incoming, element, grid, true)
-    edge.path = path
-    edges.push(edge)
-  }
-  return edges
-}
