+++ lib/utils/elementUtils.js
@@ -106,3 +106,13 @@ export function isStartIntermediate(element) {
   return (is(element, 'bpmn:IntermediateThrowEvent') || is(element, 'bpmn:IntermediateCatchEvent'))
       && (element.incoming === undefined || element.incoming.length === 0);
 }
+
+export function bindBoundaryEventsWithHosts (elements) {
+  const boundaryEvents = elements.filter(element => isBoundaryEvent(element));
+  boundaryEvents.forEach(boundaryEvent => {
+    const attachedTask = boundaryEvent.attachedToRef;
+    const attachers = attachedTask.attachers || [];
+    attachers.push(boundaryEvent);
+    attachedTask.attachers = attachers;
+  });
+}
