+++ test/snapshots/boundary-event.multiple.bpmn
@@ -12,8 +12,8 @@
     </bpmn2:startEvent>
     <bpmn2:sequenceFlow id="Flow_1t9f51i" sourceRef="Event_16w9732" targetRef="SubProcess_1" />
     <bpmn2:boundaryEvent id="Event_1nxutwo" attachedToRef="SubProcess_1">
-      <bpmn2:outgoing>Flow_00y4kzn</bpmn2:outgoing>
       <bpmn2:outgoing>Flow_0xhtwne</bpmn2:outgoing>
+      <bpmn2:outgoing>Flow_00y4kzn</bpmn2:outgoing>
     </bpmn2:boundaryEvent>
     <bpmn2:task id="Activity_0hwu7fi">
       <bpmn2:incoming>Flow_1p5v54y</bpmn2:incoming>
@@ -42,10 +42,10 @@
       <bpmndi:BPMNShape id="Event_1nxutwo_di" bpmnElement="Event_1nxutwo">
         <dc:Bounds x="223.66666666666669" y="92" width="36" height="36" />
       </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Activity_09a7ivi_di" bpmnElement="Activity_09a7ivi">
+      <bpmndi:BPMNShape id="Activity_17tmt4d_di" bpmnElement="Activity_17tmt4d">
         <dc:Bounds x="325" y="170" width="100" height="80" />
       </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Activity_17tmt4d_di" bpmnElement="Activity_17tmt4d">
+      <bpmndi:BPMNShape id="Activity_09a7ivi_di" bpmnElement="Activity_09a7ivi">
         <dc:Bounds x="325" y="310" width="100" height="80" />
       </bpmndi:BPMNShape>
       <bpmndi:BPMNShape id="Activity_0hwu7fi_di" bpmnElement="Activity_0hwu7fi">
@@ -60,12 +60,12 @@
         <di:waypoint x="208.33333333333334" y="490" />
         <di:waypoint x="325" y="490" />
       </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="Flow_00y4kzn_di" bpmnElement="Flow_00y4kzn">
+      <bpmndi:BPMNEdge id="Flow_0xhtwne_di" bpmnElement="Flow_0xhtwne">
         <di:waypoint x="241.66666666666669" y="128" />
         <di:waypoint x="241.66666666666669" y="350" />
         <di:waypoint x="325" y="350" />
       </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="Flow_0xhtwne_di" bpmnElement="Flow_0xhtwne">
+      <bpmndi:BPMNEdge id="Flow_00y4kzn_di" bpmnElement="Flow_00y4kzn">
         <di:waypoint x="241.66666666666669" y="128" />
         <di:waypoint x="241.66666666666669" y="210" />
         <di:waypoint x="325" y="210" />
