+++ test/snapshots/scenario.disconnected-loop.bpmn
@@ -18,21 +18,21 @@
       <bpmndi:BPMNShape id="Event_1vp4jvc_di" bpmnElement="Event_1vp4jvc">
         <dc:Bounds x="57" y="52" width="36" height="36" />
       </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="Activity_09kkekv_di" bpmnElement="Activity_09kkekv">
-        <dc:Bounds x="25" y="170" width="100" height="80" />
-      </bpmndi:BPMNShape>
       <bpmndi:BPMNShape id="Gateway_1wpffqc_di" bpmnElement="Gateway_1wpffqc" isMarkerVisible="true">
-        <dc:Bounds x="200" y="185" width="50" height="50" />
+        <dc:Bounds x="50" y="185" width="50" height="50" />
       </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="Flow_128ka86_di" bpmnElement="Flow_128ka86">
-        <di:waypoint x="125" y="210" />
-        <di:waypoint x="200" y="210" />
-      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNShape id="Activity_09kkekv_di" bpmnElement="Activity_09kkekv">
+        <dc:Bounds x="175" y="170" width="100" height="80" />
+      </bpmndi:BPMNShape>
       <bpmndi:BPMNEdge id="Flow_0qxg6ou_di" bpmnElement="Flow_0qxg6ou">
-        <di:waypoint x="225" y="235" />
+        <di:waypoint x="100" y="210" />
+        <di:waypoint x="175" y="210" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="Flow_128ka86_di" bpmnElement="Flow_128ka86">
+        <di:waypoint x="225" y="250" />
         <di:waypoint x="225" y="280" />
         <di:waypoint x="75" y="280" />
-        <di:waypoint x="75" y="250" />
+        <di:waypoint x="75" y="235" />
       </bpmndi:BPMNEdge>
     </bpmndi:BPMNPlane>
   </bpmndi:BPMNDiagram>
